refactor(swarm): pg-cleanup pass — remaining subsystems
env, domains, governance, benchmarks, csm, judges, evaluation,
research, scenarios, redteam, council, boundaries, adaptive, skills,
replay, forecaster, contracts, neurosymbolic, knowledge, detection,
chain, calibration: dedup helpers, dead code, doc/behavior alignment;
unused score_task_bundle log param dropped with its caller.


Files-Changed: 104
Scope: swarm,swarm/adaptive,swarm/benchmarks,swarm/benchmarks/allocation,swarm/benchmarks/coordination

+++ swarm/research/agents.py
@@ -684,7 +684,7 @@ class AnalysisAgent(ResearchAgent):
             std = float(np.std(values))
             ci = self._compute_ci(values)
 
-            mean_val: float = float(mean)
+            mean_val: float = mean
             claims.append(
                 Claim(
                     statement=f"Mean {metric}: {mean_val:.3f} (SD: {std:.3f})",
